+++ src/helm/proxy/clients/huggingface_tokenizer.py
@@ -78,6 +78,10 @@ class HuggingFaceTokenizers:
                     hf_tokenizer_name = "bigcode/santacoder"
                 elif tokenizer_name == "openai/clip-vit-large-patch14":
                     hf_tokenizer_name = "openai/clip-vit-large-patch14"
+                elif tokenizer_name == "Writer/palmyra-base":
+                    hf_tokenizer_name = "Writer/palmyra-base"
+                elif tokenizer_name == "bigcode/large-model":
+                    hf_tokenizer_name = "bigcode/large-model"
                 else:
                     raise ValueError(f"Unsupported HuggingFace tokenizer: {tokenizer_name}")